+++ src/Contracts/FeedFactory.php
@@ -2,7 +2,9 @@
 
 namespace Astrotomic\Stancy\Contracts;
 
+use Illuminate\Support\Collection;
+
 interface FeedFactory
 {
-    public function makeFromSheetCollectionName(string $name): array;
+    public function makeFromSheetCollectionName(string $name): Collection;
 }