+++ src/Factories/FeedFactory.php
@@ -5,6 +5,7 @@ namespace Astrotomic\Stancy\Factories;
 use Astrotomic\Stancy\Contracts\FeedFactory as FeedFactoryContract;
 use Astrotomic\Stancy\Contracts\Page as PageContract;
 use Astrotomic\Stancy\Contracts\PageFactory as PageFactoryContract;
+use Illuminate\Support\Collection;
 use Spatie\Sheets\Facades\Sheets;
 use Spatie\Sheets\Repository as SheetRepository;
 use Spatie\Sheets\Sheet;
@@ -19,17 +20,17 @@ class FeedFactory implements FeedFactoryContract
         $this->pageFactory = $pageFactory;
     }
 
-    public function makeFromSheetCollectionName(string $name): array
+    public function makeFromSheetCollectionName(string $name): Collection
     {
         $collection = Sheets::collection($name);
 
         return $this->makeFromSheetCollection($collection);
     }
 
-    protected function makeFromSheetCollection(SheetRepository $collection): array
+    protected function makeFromSheetCollection(SheetRepository $collection): Collection
     {
         return $collection->all()->map(function (Sheet $sheet): PageContract {
             return $this->pageFactory->makeFromSheet($sheet);
-        })->all();
+        });
     }
 }